[IMP] l10n_ro_invoice_report: print cash payment receipt on out_receipt
- always print the attached cash payment (chitanta) when printing an
  out_receipt document, regardless of the report variant
- add Sales Receipt title and Supplier/Customer labels for out_receipt
  in the report header
- use dedicated 'Representing counter value of sales receipt' wording
- is_bf_printed: only relevant for out_invoice
- move ResCompany to its own res_company.py file
- fix multi-company test journal search; tag basic tests post_install
- add Romanian translations for the new terms
