+++ l10n_ro_invoice_report/__manifest__.py
@@ -4,11 +4,12 @@
 {
     "name": "Romania - Invoice Report Terrabit",
     "summary": "Localizare Terrabit - Facturi, Chitanta",
-    "version": "18.0.3.4.27",
+    "version": "18.0.3.4.28",
     "author": "Dorin Hongu," "Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/l10n-romania",
     "license": "AGPL-3",
     "category": "Localization",
+    "development_status": "Mature",
     "countries": ["ro"],
     "depends": [
         "base",
