+++ l10n_ro_invoice_report/models/__init__.py
@@ -1,6 +1,8 @@
 # ©  2008-2023 Dorin Hongu <dhongu(@)gmail(.)com
 # See README.rst file on addons root folder for license details
 
+
+from . import res_company
 from . import res_partner
 from . import account_invoice
