+++ l10n_ro_invoice_report/models/account_invoice.py
@@ -101,6 +101,8 @@ class AccountInvoice(models.Model):
     def is_bf_printed(self):
         """Check if the invoice has been printed with BF (Bon Fiscal)"""
         self.ensure_one()
+        if not self.move_type == "out_invoice":
+            return False
         if hasattr(self, "receipt_print"):
             return self.receipt_print
         return False
